Move the feed reject helper to document module.
