+++ document/src/vespa/document/util/feed_reject_helper.cpp
@@ -1,15 +1,13 @@
 // Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "feed_reject_helper.h"
-#include <vespa/searchcore/proton/feedoperation/operations.h>
 #include <vespa/document/update/documentupdate.h>
 #include <vespa/document/update/assignvalueupdate.h>
 #include <vespa/document/fieldvalue/boolfieldvalue.h>
 #include <vespa/document/fieldvalue/numericfieldvalue.h>
 
-using namespace document;
 
-namespace proton {
+namespace document {
 
 bool
 FeedRejectHelper::isFixedSizeSingleValue(const document::FieldValue & fv) {
@@ -49,27 +47,6 @@ FeedRejectHelper::mustReject(const DocumentUpdate & documentUpdate) {
         }
     }
     return ! documentUpdate.getFieldPathUpdates().empty();
-}
-
-bool
-FeedRejectHelper::mustReject(const UpdateOperation & updateOperation) {
-    using namespace document;
-    if (updateOperation.getUpdate()) {
-        return mustReject(*updateOperation.getUpdate());
-    }
-    return false;
-}
-
-bool
-FeedRejectHelper::isRejectableFeedOperation(const FeedOperation & op)
-{
-    FeedOperation::Type type = op.getType();
-    if (type == FeedOperation::PUT) {
-        return true;
-    } else if (type == FeedOperation::UPDATE_42 || type == FeedOperation::UPDATE) {
-        return mustReject(dynamic_cast<const UpdateOperation &>(op));
-    }
-    return false;
 }
 
 }
